fix compilation
svn path=/trunk/kdelibs/; revision=139412

+++ khtml/html/html_formimpl.cpp
@@ -988,7 +988,7 @@ void HTMLInputElementImpl::init()
         break;
     };
     QString value = getAttribute(ATTR_VALUE).string();
-    if (m_type <= ISINDEX) {
+    if ((uint) m_type <= ISINDEX) {
         // remove newline stuff..
         QString nvalue;
         for (int i = 0; i < value.length(); ++i)
